Name graphics test files after the modules they test
Split renderer_test.py, mesh_test.py, patterns_test.py and shapes_test.py
into per-module test files beside their modules. Move the composite.py tests
out of renderer_test.py and the shapes.group_by_pattern_size tests out of
scene_test.py into files named for those modules. Fold acne_test_suite.py
into the shape tests: its name matched no pytest pattern, so its 27 tests
never ran. Rename renderer_benchmark_test.py to renderer_benchmark.py since
it benchmarks rather than tests.

pytest paz/graphics/: 285 passed, 1 skipped (GPU-only), against 267 collected
before. The delta is +27 acne tests that now run and -8 benchmark cases that
left directory collection.
