+++ paz/graphics/renderer_benchmark.py
