+++ paz/graphics/renderer_test.py
@@ -5,8 +5,7 @@ import jax
 import jax.numpy as jp
 import paz
 from paz import SE3
-from paz.graphics import composite, renderer
-from paz.graphics.shapes.sphere import intersect_canonical_sphere
+from paz.graphics import renderer
 from paz.graphics.types import (
     Material,
     PointLight,
@@ -22,24 +21,45 @@ from paz.graphics.types import (
 )
 
 
-OLD_CAMERA_POSE = jp.array(
-    [
-        [0.060758888721466064, 0.9969050884246826, 0.04988674074411392, 0.0],
-        [0.0, 0.0499790757894516, -0.9987502694129944, 0.0],
-        [-0.998152494430542, 0.060682956129312515, 0.003036673180758953, -4.242640495300293],  # fmt: skip
-        [0.0, 0.0, 0.0, 1.0],
-    ]
-)
+@pytest.fixture
+def small_image_shape():
+    return 32, 32
 
 
-NEW_CAMERA_POSE = jp.array(
-    [
-        [0.9902160167694092, 0.11968734860420227, 0.07174412161111832, 0.0],
-        [0.0, 0.5141358971595764, -0.857708752155304, 0.0],
-        [-0.13954311609268188, 0.8493169546127319, 0.5091056227684021, -4.242640495300293],  # fmt: skip
-        [0.0, 0.0, 0.0, 1.0],
-    ]
-)
+@pytest.fixture
+def camera_pose():
+    return SE3.view_transform(
+        jp.array([0.0, 0.0, 5.0]),
+        jp.array([0.0, 0.0, 0.0]),
+        jp.array([0.0, 1.0, 0.0]),
+    )
+
+
+def snapshot_path(filename):
+    return str(Path(__file__).parent / "snapshots" / filename)
+
+
+def assert_snapshot(array, filename, atol=1e-4):
+    paz.assert_snapshot(array, snapshot_path(filename), atol=atol)
+
+
+def render_scene(image_shape, camera_pose, scene, lights, mask=None,
+                 shadows=False, tiles=(1, 1), chunk_size=1024,
+                 shadow_mask=None, num_bounces=1):
+    args = image_shape, jp.pi / 3.0, camera_pose, scene, mask, lights
+    args += tiles, chunk_size
+    return renderer.render(*args, shadows, shadow_mask, num_bounces)
+
+
+def compute_max_abs_difference(array_A, array_B):
+    return float(jp.max(jp.abs(array_A - array_B)))
+
+
+def assert_render_matches(actual, expected, atol=1e-4):
+    actual_image, actual_depth = actual
+    expected_image, expected_depth = expected
+    assert compute_max_abs_difference(actual_image, expected_image) <= atol
+    assert compute_max_abs_difference(actual_depth, expected_depth) <= atol
 
 
 def build_shadow_scene():
@@ -84,349 +104,6 @@ def build_material_scene():
     cube = Cube(SE3.translation(jp.array([1.5, 0.0, 0.0])), glass_material)
     lights = [PointLight(jp.ones(3), jp.array([0.0, 10.0, 5.0]))]
     return Scene([floor, sphere, cube]), lights
-
-
-def build_legacy_rays(image_shape, y_fov, world_to_camera):
-    H, W = image_shape[:2]
-    aspect_ratio = paz.graphics.camera.compute_aspect_ratio(H, W)
-    H_world, W_world = paz.graphics.camera.compute_image_sizes(
-        y_fov, aspect_ratio
-    )
-    args = H, W, H_world, W_world
-    directions = paz.graphics.camera.build_ray_directions(*args)
-    origins = paz.graphics.camera.build_ray_origins(H, W)
-    camera_to_world = jp.linalg.inv(world_to_camera)
-    return paz.algebra.transform_rays(camera_to_world, origins, directions)
-
-
-def compute_max_abs_difference(array_A, array_B):
-    return float(jp.max(jp.abs(array_A - array_B)))
-
-
-def assert_render_matches(actual, expected, atol=1e-4):
-    actual_image, actual_depth = actual
-    expected_image, expected_depth = expected
-    assert compute_max_abs_difference(actual_image, expected_image) <= atol
-    assert compute_max_abs_difference(actual_depth, expected_depth) <= atol
-
-
-def compute_selected_shadow_depths(camera_pose, image_shape=(120, 160)):
-    scene, lights = build_shadow_scene()
-    rays = paz.graphics.camera.build_rays(image_shape, jp.pi / 3.0, camera_pose)
-    compiled = paz.graphics.scene.compile(scene, lights, None)
-    shapes, mask, lights = compiled.shapes, compiled.mask, compiled.lights
-    intersections = renderer.intersect.build_candidates(compiled, rays, None)
-    closest = renderer.intersect.find_closest(*intersections)
-    vector = lights[0].position - closest.point
-    distance = jp.squeeze(paz.algebra.compute_norms(vector, 1), axis=1)
-    light_directions = vector / jp.expand_dims(distance, 1)
-    shadow_ray_origins = renderer.shadow.compute_shadow_ray_origins(
-        closest.point, closest.normal
-    )
-    intersections = renderer.intersect.intersect_shadow_groups(
-        shapes, shadow_ray_origins, light_directions
-    )
-    hit_masks, depths, _, _, _, shape_indices = intersections
-    transparencies = jp.array([shape.material.transparency for shape in shapes])
-    shadow_masks = jp.where(jp.expand_dims(mask, 1), hit_masks, False)
-    shadow_masks = jp.where(
-        jp.expand_dims(transparencies > 0.0, 1), False, shadow_masks
-    )
-    shadow_masks, depths = renderer.shadow.select_shadow_depths(
-        shadow_masks,
-        depths,
-        shape_indices,
-        closest.primitive_index,
-        closest.normal,
-        light_directions,
-    )
-    return shadow_masks, depths, shape_indices, closest.primitive_index
-
-
-def take_shape_depths(depths, shape_indices, receiver_indices, shape_index):
-    shape_row = int(jp.argwhere(shape_indices == shape_index)[0, 0])
-    receiver_mask = receiver_indices == shape_index
-    return depths[shape_row][receiver_mask]
-
-
-# --- Helper Functions Unit Tests ---
-
-
-def test_take_closest():
-    array = jp.array(
-        [[[1, 1, 1], [2, 2, 2], [3, 3, 3]], [[4, 4, 4], [5, 5, 5], [6, 6, 6]]]
-    )
-    indices = jp.array([0, 1, 0])
-    expected = jp.array([[1, 1, 1], [5, 5, 5], [3, 3, 3]])
-    result = composite.take_closest(array, indices)
-    assert jp.array_equal(result, expected)
-
-
-def test_compute_soft_occlusion():
-    light_lengths = jp.array([10.0, 10.0, 10.0, 10.0])
-    depths = jp.array(
-        [[paz.graphics.FARAWAY, 5.0, 10.0, 15.0], [11.0, 11.0, 10.0, 11.0]]
-    )
-    rows = [[True, True, True, True], [False, False, True, False]]
-    hit_masks = jp.array(rows)
-    result = renderer.shadow.compute_soft_occlusion(
-        hit_masks, depths, light_lengths, slope=10.0
-    )
-    assert float(result[1]) > 0.9
-    assert float(result[2]) == pytest.approx(0.5)
-    assert float(result[3]) == 0.0
-    assert float(result[0]) == 0.0
-
-
-def test_compute_new_rays_reflection_is_normalized():
-    normal = jp.array([[0.0, 1.0, 0.0]])
-    eye = jp.array([[0.0, 1.0, -1.0]])
-    point = jp.array([[0.0, 0.0, 0.0]])
-    transparencies = jp.array([0.0])
-    _, direction = renderer.optics.compute_new_rays(
-        normal, eye, jp.array([1.0]), point, transparencies
-    )
-    norm = jp.linalg.norm(direction, axis=-1)
-    assert jp.allclose(norm, 1.0, atol=1e-5)
-
-
-def test_compute_new_rays_refraction_is_normalized():
-    normal = jp.array([[0.0, 0.0, -1.0]])
-    eye = jp.array([[0.0, 0.0, -1.0]])
-    point = jp.array([[0.0, 0.0, 0.0]])
-    transparencies = jp.array([1.0])
-    _, direction = renderer.optics.compute_new_rays(
-        normal, eye, jp.array([1.0 / 1.5]), point, transparencies
-    )
-    norm = jp.linalg.norm(direction, axis=-1)
-    assert jp.allclose(norm, 1.0, atol=1e-5)
-
-
-def test_compute_surface_points_offset_hit():
-    point = jp.array([[0.0, 0.0, 0.0]])
-    normal = jp.array([[0.0, 0.0, -1.0]])
-    surface_points = renderer.shadow.compute_surface_points(point, normal)
-    over_point, under_point = surface_points
-    assert over_point[0, 2] < -(renderer.shadow.SHADOW_ORIGIN_EPSILON / 2.0)
-    assert point[0, 2] > over_point[0, 2]
-    assert under_point[0, 2] > 0.0
-
-
-def test_select_shadow_depths_discard_front_side_same_shape_hits():
-    depths = jp.array(
-        [
-            [[0.2], [0.4]],
-            [[5e-4], [1000.0]],
-        ]
-    )
-    hit_masks = jp.array([[True], [True]])
-    shape_indices = jp.array([0, 1])
-    receiver_indices = jp.array([0])
-    receiver_normals = jp.array([[0.0, 1.0, 0.0]])
-    light_directions = jp.array([[0.0, 1.0, 0.0]])
-    hit_masks, depths = renderer.shadow.select_shadow_depths(
-        hit_masks,
-        depths,
-        shape_indices,
-        receiver_indices,
-        receiver_normals,
-        light_directions,
-    )
-    assert not bool(hit_masks[0, 0])
-    assert float(depths[0, 0]) == pytest.approx(paz.graphics.FARAWAY)
-    assert bool(hit_masks[1, 0])
-    assert float(depths[1, 0]) == pytest.approx(5e-4)
-
-
-def test_select_shadow_depths_keep_back_side_second_root():
-    depths = jp.array(
-        [
-            [[1e-6], [0.2]],
-            [[5e-4], [1000.0]],
-        ]
-    )
-    hit_masks = jp.array([[True], [True]])
-    shape_indices = jp.array([0, 1])
-    receiver_indices = jp.array([0])
-    receiver_normals = jp.array([[0.0, 1.0, 0.0]])
-    light_directions = jp.array([[0.0, -1.0, 0.0]])
-    hit_masks, depths = renderer.shadow.select_shadow_depths(
-        hit_masks,
-        depths,
-        shape_indices,
-        receiver_indices,
-        receiver_normals,
-        light_directions,
-    )
-    args = hit_masks, depths, jp.array([0.01])
-    result = renderer.shadow.compute_soft_occlusion(*args)
-    assert bool(hit_masks[0, 0])
-    assert float(depths[0, 0]) == pytest.approx(0.2)
-    assert bool(hit_masks[1, 0])
-    assert float(depths[1, 0]) == pytest.approx(5e-4)
-    assert float(result[0]) > 0.5
-
-
-def test_compute_scene_hit_mask():
-    hit_masks = jp.array([[False, True, False], [False, False, True]])
-    expected = jp.array([False, True, True])
-    result = composite.compute_scene_hit_mask(hit_masks)
-    assert jp.array_equal(result, expected)
-
-
-# def test_compute_occlusion_binary():
-#     norms = jp.array([10.0, 10.0])
-#     depths = jp.array([5.0, 10.0])
-#     mask = jp.array([True, True])
-#     res = renderer.compute_occlusion(norms, depths, mask)
-#     assert jp.array_equal(res, jp.array([True, False]))
-
-
-def test_select_colors():
-    depths = jp.array([[10.0, 2.0], [5.0, 8.0]])
-    colors = jp.array([[[1, 0, 0], [0, 1, 0]], [[0, 0, 1], [1, 1, 0]]])
-    expected = jp.array([[0, 0, 1], [0, 1, 0]])
-    result = composite.select_colors(depths, colors)
-    assert jp.array_equal(result, expected)
-
-
-# --- Internal Helper Function Tests ---
-
-
-def test_initialize_render_state():
-    num_rays = 10
-    rays = (jp.zeros((num_rays, 3)), jp.ones((num_rays, 3)))
-    state = renderer.rays.initialize_state(rays)
-    assert state.color.shape == (num_rays, 3)
-    assert state.throughput.shape == (num_rays, 3)
-    assert jp.all(state.refractive_index == 1.0)
-    assert state.depth.shape == (num_rays,)
-    assert state.hit_mask.dtype == bool
-
-
-def test_find_closest_intersection_args():
-    hit_masks = jp.array([[True, False], [False, True]])
-    depths = jp.array([[1.0, 10.0], [10.0, 2.0]])
-    indices = composite.find_closest_intersection_args(hit_masks, depths)
-    assert jp.array_equal(indices, jp.array([0, 1]))
-
-
-def test_gather_shape_material():
-    mat1 = Material(reflective=0.5)
-    mat2 = Material(transparency=0.8)
-    shape1 = Sphere(material=mat1)
-    shape2 = Sphere(material=mat2)
-    scene = Scene([shape1, shape2])
-    compiled = paz.graphics.scene.compile(scene, [], None)
-    indices = jp.array([0, 1])
-    material = renderer.material.gather_shape_material(compiled, indices)
-    assert material.reflectivities[0] == 0.5
-    assert material.transparencies[1] == 0.8
-    assert material.refractive_indices[0] == 1.0
-
-
-def test_gather_primitive_material_reads_mesh_fields():
-    materials = Material(
-        reflective=jp.array([0.0, 0.6]),
-        transparency=jp.array([0.0, 0.3]),
-        refractive_index=jp.array([1.0, 1.5]),
-    )
-    primitive = jp.array([1, 0, 1])
-    gather = renderer.material.gather_primitive_material
-    material = gather(materials, primitive)
-    expected_indices = jp.array([1.5, 1.0, 1.5])
-    assert jp.array_equal(material.reflectivities, jp.array([0.6, 0.0, 0.6]))
-    assert jp.array_equal(material.refractive_indices, expected_indices)
-
-
-def test_accumulate_color():
-    num_rays = 2
-    colors = jp.zeros((num_rays, 3))
-    throughput = jp.ones((num_rays, 3))
-    active_mask = jp.array([True, False])
-    intersected_colors = jp.ones((num_rays, 3))
-    reflectivities = jp.zeros((num_rays,))
-    transparencies = jp.zeros((num_rays,))
-    result = renderer.rays.accumulate_color(
-        colors,
-        throughput,
-        active_mask,
-        intersected_colors,
-        reflectivities,
-        transparencies,
-    )
-    assert jp.array_equal(result[0], jp.array([1.0, 1.0, 1.0]))
-    assert jp.array_equal(result[1], jp.array([0.0, 0.0, 0.0]))
-
-
-def test_compute_shadow_ray_origins_avoid_lit_side_self_hit():
-    points = jp.array([[0.0, 1.0, 0.0]])
-    normals = jp.array([[0.0, 1.0, 0.0]])
-    light_position = jp.array([[0.0, 3.0, -3.0]])
-    directions = paz.algebra.normalize(light_position - points)
-    origins = renderer.shadow.compute_shadow_ray_origins(points, normals)
-    hit_mask, _, _ = intersect_canonical_sphere(origins, directions)
-    assert not bool(hit_mask[0])
-
-
-# def test_compute_light_vectors():
-#     light = PointLight(
-#         jp.array([1.0, 1.0, 1.0]), position=jp.array([10.0, 0.0, 0.0])
-#     )
-#     point = jp.array([[0.0, 0.0, 0.0]])
-#     direction, distance = renderer.compute_light_vectors(light, point)
-#     assert jp.allclose(direction, jp.array([[1.0, 0.0, 0.0]]))
-#     assert jp.allclose(distance, jp.array([10.0]))
-
-
-# def test_resolve_shadow_masks():
-#     mask = jp.array([True, False])  # Object 0 casts shadow, 1 does not
-#     shadow_init = None
-#     hits = (None, None, None, None, None, jp.array([0, 1]))  # Hit indices
-#     shape_transparencies = jp.zeros(2)
-
-#     sh_masks = renderer.resolve_shadow_masks(
-#         mask,
-#         shadow_init,
-#         (jp.ones((2, 1), dtype=bool), *hits[1:]),
-#         shape_transparencies,
-#     )
-#     # Object 0 -> True & True = True. Object 1 -> True & False = False.
-#     assert sh_masks[0, 0] == True
-#     assert sh_masks[1, 0] == False
-
-
-# --- Integration Tests with Scenes ---
-
-
-@pytest.fixture
-def small_image_shape():
-    return 32, 32
-
-
-@pytest.fixture
-def camera_pose():
-    return SE3.view_transform(
-        jp.array([0.0, 0.0, 5.0]),
-        jp.array([0.0, 0.0, 0.0]),
-        jp.array([0.0, 1.0, 0.0]),
-    )
-
-
-def snapshot_path(filename):
-    return str(Path(__file__).parent / "snapshots" / filename)
-
-
-def assert_snapshot(array, filename, atol=1e-4):
-    paz.assert_snapshot(array, snapshot_path(filename), atol=atol)
-
-
-def render_scene(image_shape, camera_pose, scene, lights, mask=None,
-                 shadows=False, tiles=(1, 1), chunk_size=1024,
-                 shadow_mask=None, num_bounces=1):
-    args = image_shape, jp.pi / 3.0, camera_pose, scene, mask, lights
-    args += tiles, chunk_size
-    return renderer.render(*args, shadows, shadow_mask, num_bounces)
 
 
 def build_checkered_image(box_size=4, rows=4, cols=4):
@@ -728,34 +405,6 @@ def test_render_single_sphere_shadows_stay_local():
     assert jp.allclose(
         image_no_shadows[background_mask], image_shadows[background_mask]
     )
-
-
-def test_saved_pose_sphere_self_shadow_keeps_later_roots():
-    _, depths, shape_indices, receiver_indices = compute_selected_shadow_depths(
-        OLD_CAMERA_POSE
-    )
-    args = depths, shape_indices, receiver_indices, 0
-    sphere_depths = take_shape_depths(*args)
-    epsilon = renderer.shadow.SHADOW_SELF_HIT_EPSILON
-    assert int(jp.sum(sphere_depths < 1e-2)) > 0
-    assert float(jp.min(sphere_depths)) > epsilon
-
-
-def test_saved_pose_floor_self_hits_stay_filtered():
-    _, depths, shape_indices, receiver_indices = compute_selected_shadow_depths(
-        NEW_CAMERA_POSE
-    )
-    plane_depths = take_shape_depths(depths, shape_indices, receiver_indices, 1)
-    assert int(jp.sum(plane_depths < 1e-2)) == 0
-
-
-@pytest.mark.skipif(jax.default_backend() != "gpu", reason="GPU only")
-def test_gpu_saved_pose_floor_has_no_near_zero_self_hits():
-    _, depths, shape_indices, receiver_indices = compute_selected_shadow_depths(
-        NEW_CAMERA_POSE, image_shape=(240, 320)
-    )
-    plane_depths = take_shape_depths(depths, shape_indices, receiver_indices, 1)
-    assert int(jp.sum(plane_depths < 1e-2)) == 0
 
 
 def test_render_shadow_mask(small_image_shape, camera_pose):
