+++ paz/graphics/scene_test.py
@@ -55,87 +55,6 @@ def shape_512(material, pattern_512):
     )
 
 
-@pytest.fixture
-def shape_default(material):
-    """Provides a shape that uses the default 1x1 pattern."""
-    return Shape(
-        transform=jp.eye(4),
-        type=paz.graphics.CONE,
-        material=material,
-        pattern=Pattern(),
-    )
-
-
-def test_group_by_size_with_empty_list():
-    """Tests that an empty list of shapes results in an empty dictionary."""
-    shapes_list = []
-    result = paz.graphics.shapes.group_by_pattern_size(shapes_list)
-    assert result == {}
-
-
-def test_group_by_size_with_homogenous_list(shape_256_A, shape_256_B):
-    """Tests grouping a list where all shapes have the same image size."""
-    shapes_list = [shape_256_A, shape_256_B]
-    result = paz.graphics.shapes.group_by_pattern_size(shapes_list)
-
-    assert len(result) == 1
-    assert (256, 256) in result
-
-    result_list = result[(256, 256)]
-    assert len(result_list) == 2
-
-    # FIX: Use `is` to check for object identity, avoiding the ValueError.
-    assert result_list[0] is shape_256_A
-    assert result_list[1] is shape_256_B
-
-
-def test_group_by_size_with_heterogeneous_list(shape_256_A, shape_512):
-    """Tests grouping a list with two different image sizes."""
-    shapes_list = [shape_256_A, shape_512]
-    result = paz.graphics.shapes.group_by_pattern_size(shapes_list)
-
-    assert len(result) == 2
-    assert (256, 256) in result
-    assert (512, 512) in result
-    assert result[(256, 256)] == [shape_256_A]
-    assert result[(512, 512)] == [shape_512]
-
-
-def test_group_by_size_with_default_pattern(shape_default):
-    """shapes with default patterns are grouped correctly by size (1, 1)."""
-    shapes_list = [shape_default]
-    result = paz.graphics.shapes.group_by_pattern_size(shapes_list)
-
-    assert len(result) == 1
-    assert (1, 1) in result
-    assert result[(1, 1)] == [shape_default]
-
-
-def test_group_by_size_with_complex_mixed_list(
-    shape_256_A, shape_512, shape_default, shape_256_B
-):
-    """Tests a complex mix of shapes with different and shared image sizes."""
-    shapes_list = [shape_256_A, shape_512, shape_default, shape_256_B]
-    result = paz.graphics.shapes.group_by_pattern_size(shapes_list)
-
-    assert len(result) == 3
-    assert (256, 256) in result
-    assert (512, 512) in result
-    assert (1, 1) in result
-
-    # Check the contents of each group
-    group_256 = result[(256, 256)]
-    assert len(group_256) == 2
-    assert len(result[(512, 512)]) == 1
-    assert len(result[(1, 1)]) == 1
-
-    # FIX: Use `is` to check for object identity. The order is preserved.
-    assert group_256[0] is shape_256_A
-    assert group_256[1] is shape_256_B
-    assert result[(512, 512)][0] is shape_512
-    assert result[(1, 1)][0] is shape_default
-
-
 def test_compute_bounces_default(shape_256_A):
     """Tests that compute_bounces returns 1 for default materials."""
     shapes = [shape_256_A]
